Merge refund plus develop

See merge request kkm_android/ofdapp!40

+++ data/src/main/java/kg/nurtelecom/data/history_by_id/Result.kt
@@ -13,5 +13,5 @@ data class Result (
 	val nspType : String,
 	val alreadyPayed : Int,
 	val receipt : Receipt,
-	val debt : Int
+	val debt : Double
 )
